% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTMDR.R
\name{PTMDR}
\alias{PTMDR}
\title{PTMDR}
\usage{
PTMDR(bfile, epii = NULL, epiall, snp, snplist = NULL,
  plink = "plink", out = ".")
}
\arguments{
\item{bfile}{The bfile variable is required, the user needs to define the path and name of the Plink bfile (bed, bim and fam) without the extension, the three files must be named in the same way. For example genotype.bed, genotype.bim, genotype.fam the variable bfile would be (bfile = "genotype").}

\item{epiall}{The epiall variable is required, the user needs to define the path and name of the epi file generated by VARI3 usually episelecrjoint.epi.cc.}

\item{snp}{The snp variable is required. It refers to the SNP that we want to find high order interactions and convert to MDR.}

\item{plink}{The plink variable is by default "plink" but it is necessary to say the path of the previously downloaded plink file if it has not been added to the system executable path.}

\item{out}{The out variable determines the output directory where all files generated by PTMDR will be written, ptmdr.txt will be the input file to MDR analysis.}
}
\description{
Plink To Multifactor Dimensionality Reduction
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya.
Help documentation has generated by roxygen2.
}
