% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLTO.R
\name{TLTO}
\alias{TLTO}
\title{TLTO}
\usage{
TLTO(l1, l2, bfile, out = ".", core = detectCores() - 1, plink = "plink")
}
\arguments{
\item{l1}{The l1 variable is required, the user have to specify the SNPs position. For example, l1 = "5:141025162".}

\item{l2}{The l2 variable is required, the user have to specify the SNPs position. For example, l2 = "4:90637601".}

\item{bfile}{The bfile variable is required, the user needs to define the path and name of the Plink bfile (bed, bim and fam) without the extension, the three files must be named in the same way. For example genotype.bed, genotype.bim, genotype.fam the variable bfile would be (bfile = "genotype").}

\item{out}{The out variable determines the output directory where all files generated by TLTO will be written.}

\item{core}{The core variable is by default core = detectCores() -1. It refers to the number of cores used.}

\item{plink}{The plink variable is by default "plink" but it is necessary to say the path of the previously downloaded plink file if it has not been added to the system executable path.}
}
\description{
Two Locus To Odd
}
\details{
This function has developed by Alejandro Cisterna Garcia as part of his PhD mentored by Juan Antonio Botía Blaya.
Help documentation has generated by roxygen2.
}
