% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARI3.R
\name{VARI3}
\alias{VARI3}
\title{VARI3}
\usage{
VARI3(
  bfile,
  out = ".",
  primarylist = NULL,
  genelist = NULL,
  plink = "plink",
  core = detectCores() - 1,
  varpl = "annotate_variation.pl",
  ANNOVAR = T,
  clump = T,
  build = "hg19",
  db = "../../Doctorado/Annovar/annovar.latest/annovar/humandb/",
  Wu = T,
  covar = NULL,
  ncov = "1-2",
  AsT = 1e-04,
  MAF = 0.01
)
}
\arguments{
\item{bfile}{The bfile variable is required, the user needs to define the path and name of the Plink bfile (bed, bim and fam) without the extension, the three files must be named in the same way. For example genotype.bed, genotype.bim, genotype.fam the variable bfile would be (bfile = "genotype").}

\item{out}{The out variable determines the output directory where all files generated by VARI3 will be written.}

\item{primarylist}{prymarylist refers to the file with the list of genes where the selection of primary SNPs is searched, the file must not have a header.}

\item{genelist}{genelist refers to the file with the list of genes where you want to look for the epistasis, the secondary SNPs, the file must not have header.}

\item{plink}{The plink variable is by default "plink" but it is necessary to say the path of the previously downloaded plink file if it has not been added to the system executable path.}

\item{core}{The core variable is by default core = detectCores() -1. It refers to the number of cores used.}

\item{varpl}{The varpl variable is by default "annotate_variation.pl", but it is necessary to say the path of the .pl file if it has not been added to the system executable path.}

\item{ANNOVAR}{The ANNOVAR variable is by default ANNOVAR = T, for the annotation with ANNOVAR. Note, at the moment the functionality with false is not implemented.}

\item{clump}{The clump variable is by default clump = T, to perform clumping. We recommend clumping if the analysis is for more than 2000 people.}

\item{build}{The variable build is used to determine the version of the reference genome, it is by default (build = "hg19").}

\item{db}{The variable db is necessary to define where annovar reference databases are located, normally in the downloaded file they follow the path "annovar/humandb/".}

\item{Wu}{The variable Wu refers to the proposed correction for epistasis tests, by default it is Wu = T.}

\item{covar}{The covar variable refers to the file where the covariates are located. For example, (covar = "covarfile.cov"). By default it is null.}

\item{ncov}{The ncov variable refers to how many covariates are selected from the covariables file, they are selected depending on the column we want, as in plink. For example to select the first four of a file (ncov = "1-4")}

\item{AsT}{The AsT variable refers to the p.maximum value for selecting SNPs in the association analysis and is by default (AsT = 0.0001).}

\item{MAF}{The MAF variable refers to the MAF from which SNPs are selected for analysis, by default (MAF = 0.01) to avoid very rare SNPs.}
}
\description{
Approach to VARIant-VARIant interaction through VARIable thresholds and hypothesis testing. VARI3 automates the selection and analysis of the most promising SNPs to identify epistasis.
}
\details{
This function has developed by Alejandro Cisterna Garcia as part of his PhD mentored by Juan Antonio Botía Blaya.
Help documentation has generated by roxygen2.
}
