% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSNP3.R
\name{VSNP3}
\alias{VSNP3}
\title{VSNP3}
\usage{
VSNP3(
  bfile,
  out = ".",
  snplist = NULL,
  plink = "plink",
  core = detectCores() - 1,
  varpl = "annotate_variation.pl",
  ANNOVAR = F,
  build = "hg19",
  db = "../../Doctorado/Annovar/annovar.latest/annovar/humandb/"
)
}
\arguments{
\item{bfile}{The bfile variable is required, the user needs to define the path and name of the Plink bfile (bed, bim and fam) without the extension, the three files must be named in the same way. For example genotype.bed, genotype.bim, genotype.fam the variable bfile would be (bfile = "genotype").}

\item{out}{The out variable determines the output directory where all files generated by VARI3 will be written.}

\item{snplist}{snplist refers to the file with the list of SNPs where we are going to look for epistasis}

\item{plink}{The plink variable is by default "plink" but it is necessary to say the path of the previously downloaded plink file if it has not been added to the system executable path.}

\item{core}{The core variable is by default core = detectCores() -1. It refers to the number of cores used.}

\item{varpl}{The varpl variable is by default "annotate_variation.pl", but it is necessary to say the path of the .pl file if it has not been added to the system executable path.}

\item{ANNOVAR}{The ANNOVAR variable is by default ANNOVAR = T, for the annotation with ANNOVAR. Note, at the moment the functionality with false is not implemented.}

\item{build}{The variable build is used to determine the version of the reference genome, it is by default (build = "hg19").}

\item{db}{The variable db is necessary to define where annovar reference databases are located, normally in the downloaded file they follow the path "annovar/humandb/".}
}
\description{
Approach to VARIant-VARIant interaction through VARIable thresholds and hypothesis testing. VARI3 automates the selection and analysis of the most promising SNPs to identify epistasis.
}
\details{
This function has developed by Alejandro Cisterna García as part of his PhD mentored by Juan Antonio Botía Blaya.
Help documentation has generated by roxygen2.
}
